+++ lib/native_packages/support.rb
@@ -12,7 +12,7 @@ require "tmpdir"
 require "yaml"
 
 module NativePackages
-  VERSION = "0.3.0"
+  VERSION = "0.3.1"
   class Error < StandardError; end
 
   module Support